Removed postA from pairing method
PostAlpha removed as it is no longer used in peak pairing algorithm.  This parameter was left over from when pre-filtration was being used in the pairing method.  Since then the pre-filtration is now done post pairing as part of an optional validation/data-cleaning step and has been moved outside the pairing algorithm.  This optional data filtration now uses postA.