Partial Fix of nC13 error -- was changed to nLabel
Discovered that nC13 was not changed to nLabel in most scripts.  This is likely an oversite from when  the package was converted to be able to do catabolism testing.  This is when nC13 was changed to nLabel in groupmzrange but this was not refelcted across the other scripts.

I have only changed the core scripts so far, others will need to be changed for full compatibility