Bumped to coleXcmsMethods2  nLabel Implemented
This is a bare bones implementation of the nLabel and mLabel parameters.  All core functionality is working, but down stream scripts such as OPLSDA, database searching, CAMERA, ect are now broken.  Also, the nLabel and mLabel parameter only work for a "main" analysis.  They are not supported by additional validation methods (like tTest_validation) or any of the heuristic methods.

Prepended old_ to previous methods. Removed group_nlabel_validationOutput as it is now the main group_validationOutput