Started Single Genotype Rework -- Fixed peak pairing error
I discovered that when running with just one genotype, the first error you hit is an error about an innapropriate index on now line 95 of tFilterPreProcess.R

This was due to useing mat.or.vec() to construct tTests on line 48.  This caused it to be a vector instead of a matrix, causing the indexing to fail.  This was fixed by switching to matrix() instead, ensuring a matrix was produced even when only one genotype was present.

The rest of the tFilterPreProcess routine and the labeledPeaks_GroupPairing_tTest routine that calles tFilter both now run to completition.

I still need to ensure this produces the expected results, and that it does not break anything when running with more than one genotype.