Remove hardcoding of ppm-based expansion
The use of ppm based expansion was hardcoded into group_printEICs and group_validationOutput.

Now, if ppm is NULL, the old/default method from the xcms package will be used for calculating the M/z bounds for EIC generation instead of the ppm based approach.

Will need to document the ppm=NULL to switch method behaviour.  This will only happen if the user explicetly sets it to NULL, as the default is ppm=50 (I may want to turn this down to something more concervative like 20 or 25)